Consolidate the delegation tool face and workflow-output periphery
Inline the delivery-summary collector into the strategy closure and
delete its file, factory, and single-implementation interface. Merge
the four delegation availability micro-files into one
delegationAvailability module, fold assertWorkflowFilesExist into
inputFields, and collapse the scoped/unscoped agent resolvers into
getDelegationAgent(s).

Boundary shims: persistChildRunTurnState and getOutputFileName are
inlined at their callers, and persistence's onSnapshot re-parse
override is deleted — the durable sink already validates via
ExecutionMetaSchema on write. Delete the runner-side
dependency-fingerprint re-verification; the engine's
refreshDependencyIdentity is the single owner of attempt identity and
journal keys are unchanged.

Fix a live bug in XML packing: packAdditionalXmlFiles hand-built the
flat stem with the raw model name, so flat outputs of dotted models
(gpt-4.5 era) never packed; both candidates now compose from the SSOT
stem helpers, pinned by a regression test. Rename the misnamed
legacyWorkflowOutputStem to workflowOutputCopyStem — it names files
the extension's Save-as-copy action writes today.

Home the skip/retry control vocabulary in @shared/schemas beside
WORKFLOW_CALL_STATUS; the CLI's host-local mirror and the registry's
unused map key are deleted.
