+++ packages/extension/src/commands/latex/compareCommands.ts
@@ -22,7 +22,7 @@ import {
 import * as logger from '@logger/logUtils';
 import type { AcceptCopyMeta, FileLocation } from '@shared/schemas';
 import { DIFF_REGISTRATION_DELAY_MS } from '@shared/constants/latex';
-import { legacyWorkflowOutputStem } from '@shared/constants/workflowOutput';
+import { workflowOutputCopyStem } from '@shared/constants/workflowOutput';
 import { AbsoluteFS } from '@utils/files';
 import { toErrorMessage } from '@utils/errors/errorMessage';
 
@@ -168,7 +168,7 @@ function buildCopyTarget(
   copyMeta: AcceptCopyMeta,
 ): ReplaceOrCopyTarget {
   const parsed = path.parse(baseLocation.absolutePath);
-  const stem = legacyWorkflowOutputStem({
+  const stem = workflowOutputCopyStem({
     base: parsed.name,
     agent: copyMeta.agent,
     model: copyMeta.model,
