+++ src/agent/runtime/agentToolResolution.ts
@@ -41,13 +41,11 @@ import {
 } from '@tools/diagnosticsRuntimeCapabilities';
 import {
   availableModelNamesFromOptions,
-  withDelegationModelAvailability,
-} from '@tools/delegation/delegationModelAvailability';
-import {
   visibleDelegationAgentsBlock,
   withDelegationAgentAvailability,
-} from '@tools/delegation/delegationAgentAvailability';
-import { withDelegationWorktreeAvailability } from '@tools/delegation/delegationWorktreeAvailability';
+  withDelegationModelAvailability,
+  withDelegationWorktreeAvailability,
+} from '@tools/delegation/delegationAvailability';
 import { toErrorMessage } from '@utils/errors/errorMessage';
 import {
   SharedToolInjectionRegistry,
