+++ src/agent/workflowScript/index.ts
@@ -13,7 +13,6 @@ export type {
   WorkflowAgentCallOptions,
   WorkflowAgentInvocation,
   WorkflowAgentRunner,
-  WorkflowControlAction,
   WorkflowJournalEntry,
   WorkflowScriptControl,
   WorkflowScriptEvent,
