+++ src/shared/constants/workflowOutput.ts
@@ -5,9 +5,11 @@
  * Per-execution isolation (executions/{id}/...) provides uniqueness;
  * agent/model/round-in-basename tokens are no longer needed.
  *
- * Filename-era compatibility helpers (the legacy grammar below) require
- * agent-name parsing and are still consumed by workspace migration readers
- * and one copy writer.
+ * The `<base>_<agent>_r{round}_<model>` grammar below is not purely historic:
+ * workspace migration readers still match files pre-refactor runs wrote with
+ * it, and the extension's "Save as copy" action still writes new files with
+ * it today. Everything in it derives from one place so a reader can never
+ * spell a name a writer would not produce.
  */
 
 // Third-party imports
@@ -36,7 +38,15 @@ export function workflowOutputRoundDir(round: number): string {
   return `r${round}`;
 }
 
-/** Build a runDir-relative workflow output path for a round. */
+/**
+ * Build a runDir-relative workflow output path for a round: `r{round}/output.{ext}`.
+ *
+ * IMPORTANT: callers MUST resolve this through a TaskRunFileService bound to an
+ * executionId. The fixed-stem filename is only collision-safe when combined
+ * with per-execution run storage; a workspace-scoped resolution would route
+ * every round to the same `<workspace>/r{round}/output.{ext}` and clobber
+ * outputs across runs.
+ */
 export function workflowOutputPath(params: {
   ext: string;
   round: number;
@@ -45,21 +55,22 @@ export function workflowOutputPath(params: {
 }
 
 // ============================================================================
-// Filename-era workflow output compatibility grammar
+// Agent/round/model filename grammar
 // ============================================================================
 //
 // Before workflow outputs moved to execution-scoped `r{round}/output.*`
 // paths, their agent, round, and model were encoded in workspace filenames.
 // Housekeeping, XML packing, latexdiff discovery, and the extension's
-// "Save as copy" action still consume this grammar.
+// "Save as copy" action still consume this grammar — the last of those still
+// writes it.
 
 /** Normalize a model name to the filename-era form (dots stripped). */
 export function normalizeLegacyModel(model: string): string {
   return model.replaceAll('.', '');
 }
 
-/** First-name chunk used in pre-refactor filenames. */
-export function getAgentFirstNameChunk(agent: string): string {
+/** First-name chunk used in the `<base>_<chunk>_r{round}_<model>` grammar. */
+function getAgentFirstNameChunk(agent: string): string {
   const cleanAgent = getCleanAgentName(agent);
   if (cleanAgent.startsWith('write-')) {
     return cleanAgent.split('-')[1];
@@ -70,8 +81,12 @@ export function getAgentFirstNameChunk(agent: string): string {
   return cleanAgent.split('-')[0];
 }
 
-/** Filename-era stem: `<base>_<chunk>_r{round}_<normalizedModel>`. */
-export function legacyWorkflowOutputStem(params: {
+/**
+ * The `<base>_<chunk>_r{round}_<normalizedModel>` stem: the name filename-era
+ * runs wrote, and the one "Save as copy" still writes beside a base file
+ * today. Every reader of that layout composes its names from here.
+ */
+export function workflowOutputCopyStem(params: {
   base: string;
   agent: string;
   model: string;
