+++ src/test-kernel/agent/runtime/ChildRunLoop.vitest.ts
@@ -16,7 +16,6 @@ const mocks = vi.hoisted(() => ({
   finalizeExecution: vi.fn(),
   persistChildRunReport: vi.fn(),
   persistChildRunResultMeta: vi.fn(),
-  persistChildRunTurnState: vi.fn(),
   deliverChildRunFollowUp: vi.fn(),
   runWithOwnedExecutionLease: vi.fn(
     (_executionId: ExecutionId, operation: () => unknown) => operation(),
@@ -24,6 +23,8 @@ const mocks = vi.hoisted(() => ({
   leaseLossListener: undefined as (() => void) | undefined,
 }));
 
+// Turn-state persistence runs against the real (memfs-backed) execution store:
+// the loop writes it best-effort and no assertion here depends on it.
 vi.mock('@agent/storage', async (importOriginal) => ({
   ...(await importOriginal<typeof import('@agent/storage')>()),
   finalizeExecution: mocks.finalizeExecution,
@@ -52,7 +53,6 @@ vi.mock('@agent/runtime/executionOwnership', () => ({
 vi.mock('@tools/delegation/childRunDelivery', () => ({
   persistChildRunReport: mocks.persistChildRunReport,
   persistChildRunResultMeta: mocks.persistChildRunResultMeta,
-  persistChildRunTurnState: mocks.persistChildRunTurnState,
   deliverChildRunFollowUp: mocks.deliverChildRunFollowUp,
 }));
 
@@ -265,7 +265,6 @@ beforeEach(() => {
     return { kind: 'persisted' as const, msg };
   });
   mocks.persistChildRunResultMeta.mockResolvedValue({ kind: 'skipped' });
-  mocks.persistChildRunTurnState.mockResolvedValue({ kind: 'persisted' });
   mocks.deliverChildRunFollowUp.mockResolvedValue({ kind: 'delivered' });
 });
