+++ src/tools/delegation/childRunDelivery.ts
@@ -1,9 +1,5 @@
 /** Shared persistence and parent-continuation delivery for child runs. */
-import {
-  type ChildTurnState,
-  getExecutionStore,
-  type ResultMeta,
-} from '@agent/storage';
+import { getExecutionStore, type ResultMeta } from '@agent/storage';
 import type { SessionHandle } from '@agent/runtime/SessionHandle';
 import { submitFollowUp } from '@agent/followUp/ToolUseFollowUp';
 import type { FollowUpQueueInput } from '@agent/followUp/FollowUpQueue';
@@ -35,24 +31,6 @@ export async function persistChildRunResultMeta(
 ): Promise<ChildRunReportResult> {
   try {
     await getExecutionStore(executionId).writeResultMeta(resultMeta);
-    return { kind: 'persisted' };
-  } catch (err) {
-    return { kind: 'failed', err };
-  }
-}
-
-/**
- * Persist turn attribution for the report/result slots (#9531). Unlike the
- * report and manifest, this is attribution metadata: a failure downgrades
- * /report//result turn labeling but never the delivered result itself, so it
- * does not mark the lease undurable.
- */
-export async function persistChildRunTurnState(
-  executionId: ExecutionId,
-  state: ChildTurnState,
-): Promise<ChildRunReportResult> {
-  try {
-    await getExecutionStore(executionId).writeTurnState(state);
     return { kind: 'persisted' };
   } catch (err) {
     return { kind: 'failed', err };
