+++ src/tools/delegation/workflowScriptAgentRunner.ts
@@ -29,12 +29,14 @@ import {
   executeStableSubagentInBand,
   SubagentDurabilityError,
 } from './inBandSubagentExecution';
-import { rejectOversizedBibAttachments } from './inputFields';
+import {
+  assertWorkflowFilesExist,
+  rejectOversizedBibAttachments,
+} from './inputFields';
 import {
   requireVisibleAgent,
   selectAvailableDelegationModel,
 } from './proposalFlow';
-import { assertWorkflowFilesExist } from './workflowFileValidation';
 
 async function resolveInvocationFileList(
   parentExecutionId: LaunchRunContext['runScope']['executionId'],
@@ -177,29 +179,6 @@ export function createWorkflowScriptAgentRunner(
   const { runScope } = parent;
 
   return async (invocation) => {
-    if (invocation.dependencyFingerprint !== undefined) {
-      let launchFingerprint: string;
-      try {
-        launchFingerprint = await fingerprintWorkflowAgentDependencies(
-          run.executionId,
-          invocation.options,
-        );
-      } catch (error) {
-        if (error instanceof WorkflowRunAbortError) throw error;
-        throw new WorkflowRunAbortError(
-          formatError(
-            'Workflow agent() file dependencies could not be fingerprinted while the child was launching',
-            error,
-          ),
-          { cause: error },
-        );
-      }
-      if (launchFingerprint !== invocation.dependencyFingerprint) {
-        throw new WorkflowRunAbortError(
-          'Workflow agent() file dependencies changed while the child was launching; rerun the saved workflow script.',
-        );
-      }
-    }
     const logicalExecutionId = deriveExecutionId({
       checkpointId,
       key: invocation.key,
